created and moved the imperative API logic to a hook
