+++ packages/bruno-app/src/components/CodeMirrorSearch/index.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef, useCallback, forwardRef, useImperativeHandle } from 'react';
+import React, { useState, useEffect, useRef, useCallback, forwardRef } from 'react';
 import { IconRegex, IconArrowUp, IconArrowDown, IconX, IconLetterCase, IconLetterW, IconChevronRight, IconReplace, IconArrowsExchange2 } from '@tabler/icons';
 import ToolHint from 'components/ToolHint';
 import StyledWrapper from './StyledWrapper';
@@ -6,6 +6,7 @@ import useDebounce from 'hooks/useDebounce';
 import { replaceSingle, replaceAll } from './replaceUtils';
 import { findSearchMatches, createCacheKey } from './searchUtils';
 import { markViewportMatches, clearMarks } from './markingUtils';
+import { useSearchBarHandle } from './useSearchBarHandle';
 
 const CodeMirrorSearch = forwardRef(({ visible, editor, readOnly, onClose }, ref) => {
   const [searchText, setSearchText] = useState('');
@@ -126,73 +127,26 @@ const CodeMirrorSearch = forwardRef(({ visible, editor, readOnly, onClose }, ref
     }
   }, [editor, onClose]);
 
-  useImperativeHandle(ref, () => ({
-    focus: () => {
-      if (inputRef.current) {
-        inputRef.current.focus();
-      }
-    },
-    setSearch: (text, cursorPos) => {
-      setSearchText(text);
-      if (cursorPos && editor && text) {
-        const matches = findSearchMatches(editor, text, regex, caseSensitive, wholeWord);
-        const startsAtOrAfterCursor = (match) =>
-          match.from.line > cursorPos.line
-          || (match.from.line === cursorPos.line && match.from.ch >= cursorPos.ch);
-
-        const matchAtCursorIdx = matches.findIndex(startsAtOrAfterCursor);
-        const targetIdx = matchAtCursorIdx >= 0 ? matchAtCursorIdx : 0;
-        // Pre-populate cache so doSearch hits it regardless of path
-        searchMatches.current = matches;
-        searchCacheKey.current = createCacheKey(docVersion.current, text, regex, caseSensitive, wholeWord);
-        // Set both count and index immediately to avoid "1 / N" flash before debounce fires
-        setMatchCount(matches.length);
-        setMatchIndex(targetIdx);
-        doSearch(text, targetIdx, null, true);
-        initialIndexRef.current = { idx: targetIdx, forText: text };
-      } else {
-        setMatchIndex(0);
-      }
-      setTimeout(() => {
-        if (inputRef.current) {
-          inputRef.current.focus();
-          inputRef.current.select();
-        }
-      }, 0);
-    },
-    focusAtCursor: (cursorPos) => {
-      // If there's existing search text, navigate to nearest match at cursor
-      if (cursorPos && editor && searchText) {
-        const matches = searchMatches.current.length
-          ? searchMatches.current
-          : findSearchMatches(editor, searchText, regex, caseSensitive, wholeWord);
-        const startsAtOrAfterCursor = (match) =>
-          match.from.line > cursorPos.line
-          || (match.from.line === cursorPos.line && match.from.ch >= cursorPos.ch);
-        const targetIdx = matches.findIndex(startsAtOrAfterCursor);
-        const resolvedIdx = targetIdx >= 0 ? targetIdx : 0;
-        searchMatches.current = matches;
-        searchCacheKey.current = createCacheKey(docVersion.current, searchText, regex, caseSensitive, wholeWord);
-        setMatchCount(matches.length);
-        setMatchIndex(resolvedIdx);
-        initialIndexRef.current = { idx: resolvedIdx, forText: searchText };
-        doSearch(searchText, resolvedIdx);
-      }
-      setTimeout(() => {
-        if (inputRef.current) {
-          inputRef.current.focus();
-          inputRef.current.select();
-        }
-      }, 0);
-    },
-    openReplace: () => {
-      setReplaceVisible(true);
-      setTimeout(() => replaceInputRef.current?.focus(), 0);
-    },
-    close: () => {
-      handleSearchBarClose();
-    }
-  }));
+  useSearchBarHandle({
+    ref,
+    editor,
+    searchText,
+    regex,
+    caseSensitive,
+    wholeWord,
+    searchMatches,
+    searchCacheKey,
+    docVersion,
+    initialIndexRef,
+    inputRef,
+    replaceInputRef,
+    setSearchText,
+    setMatchCount,
+    setMatchIndex,
+    setReplaceVisible,
+    doSearch,
+    handleSearchBarClose
+  });
 
   useEffect(() => {
     // If setSearch/focusAtCursor pre-populated an index for a specific text but the debounce
@@ -271,7 +225,6 @@ const CodeMirrorSearch = forwardRef(({ visible, editor, readOnly, onClose }, ref
 
   const handleSearchTextChange = (text) => {
     setSearchText(text);
-    // setMatchIndex(0);
   };
 
   const handleToggleRegex = () => {
